requirements.txt
streamlit
openai
pdfplumber
pdf2image
serpapi
pillow
google-search-results
easyocr
python-dotenv